[Added] `pkg` CLI
A few key points of interest:

- Utilities that are relevant to both development and running of Hexa are moved to the `pkg` CLI, away from the `dev` CLI.
- Attempts have been made to make package setup tasks runnable when bootstrapping (e.g., first run as `./bin/kg setup` after cloning the repo).

Note that the "setup" task assumes that the consumer is comfortable with the script/CLI installing a fair bit of software such as:

  - brew
  - asdf
  - direnv

These and more provide mechanisms for installing software required for running Hexa, and some for improving the workspace ergonomics. An example: once that setup is run, `pkg serve` may be run from within any directory in the project codebase.

Signed-off-by: Corey Innis <coreyti@me.com>